Extracted out ViewWithSidebar component
